test(optimize): compare old floodfill with dino floodfill
It seems dino floodfill is around twice as fast for complex and larger
bounds and it is -30% to 100% faster for small and less complex bounds.
However, dino floodfill seems to be twice as faster than old floodfill
on average.

Also the number of iterations in old floodfill is 4 times than that
of dino floodfill.