perf(optimize): avoid populating mask matrix outside the floodfill bounds
No need to populate the mask matrix for the entire player bounds, instead
populate only the area that is being floodfilled (closedBounds). This
improves peformance greatly for large areas.

However, we can no longer track the bounds of lost area, but it's
insignificant for the performance gain as the only advantage for
tracking lost area is to clearAllPlayerTiles().

Also the visualisation of the player bounds will be inaccurate from
now on, so for visual debugging purposes, use the commits before this.